% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotdot-package.R
\docType{package}
\name{dotdot-package}
\alias{dotdot}
\alias{dotdot-package}
\title{dotdot: Enhanced Assignment to Overwrite or Grow Objects}
\description{
Use `..` on the right hand side of the `:=` operator as a shorthand for the 
    left hand side, so that `var := f(..) + ..` is equivalent to
    `var <- f(var) + var`. This permits the user to be explicit about growing an
    object or overwriting it using its previous value, avoids repeating a 
    variable name, and saves keystrokes, time, visual space and cognitive load.
}
\keyword{internal}
