% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotdot.R
\name{dotdot_first}
\alias{dotdot_first}
\title{Make sure `:=` is unmasked}
\usage{
dotdot_first()
}
\description{
This is not necessary to call this function after attaching
*rlang*/*tidyverse* or *data.table*
}
