% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotdot.R
\name{:=}
\alias{:=}
\title{Enhanced assignment}
\usage{
":="(e1, e2)
}
\description{
Use \code{..} on the right hand side as a shorthand for the left hand side
}
\examples{
x <- factor(letters[1:3])
levels(x) := c(.., "level4")
x
}
